issue #112 refactor: BaseViewModel/UiState/Event/Effect 를 :core:common 으로 이동
MVI 베이스(BaseViewModel, UiState/UiEvent/UiEffect, *_EFFECTS_KEY 상수)를
:core:common 으로 이동한다. ui 패키지의 다른 심볼과 섞이지 않도록
심볼 단위로 import 를 갱신했다.
