+++ app/build.gradle.kts
@@ -47,6 +47,7 @@ fun getLocalProperty(key: String): String = gradleLocalProperties(rootDir, provi
 dependencies {
     implementation(project(":core:designsystem"))
     implementation(project(":core:model"))
+    implementation(project(":core:common"))
 
     implementation(libs.core.ktx)
     implementation(libs.activity.compose)
