+++ app/src/main/java/cord/eoeo/momentwo/ui/albumdetail/subalbumlist/SubAlbumListScreen.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.LifecycleStartEffect
 import coil.ImageLoader
-import cord.eoeo.momentwo.ui.START_EFFECTS_KEY
+import cord.eoeo.momentwo.core.common.START_EFFECTS_KEY
 import cord.eoeo.momentwo.core.model.SubAlbumItem
 
 @OptIn(ExperimentalFoundationApi::class)
