+++ app/src/main/java/cord/eoeo/momentwo/ui/createalbum/CreateAlbumContract.kt
@@ -1,8 +1,8 @@
 package cord.eoeo.momentwo.ui.createalbum
 
-import cord.eoeo.momentwo.ui.UiEffect
-import cord.eoeo.momentwo.ui.UiEvent
-import cord.eoeo.momentwo.ui.UiState
+import cord.eoeo.momentwo.core.common.UiEffect
+import cord.eoeo.momentwo.core.common.UiEvent
+import cord.eoeo.momentwo.core.common.UiState
 import cord.eoeo.momentwo.core.model.FriendItem
 
 class CreateAlbumContract {
