+++ app/src/main/java/cord/eoeo/momentwo/ui/createalbum/CreateAlbumViewModel.kt
@@ -4,7 +4,7 @@ import android.util.Log
 import androidx.lifecycle.viewModelScope
 import cord.eoeo.momentwo.domain.album.RequestCreateAlbumUseCase
 import cord.eoeo.momentwo.domain.friend.GetFriendListUseCase
-import cord.eoeo.momentwo.ui.BaseViewModel
+import cord.eoeo.momentwo.core.common.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
