+++ app/src/main/java/cord/eoeo/momentwo/ui/friend/FriendViewModel.kt
@@ -3,7 +3,7 @@ package cord.eoeo.momentwo.ui.friend
 import android.util.Log
 import androidx.lifecycle.viewModelScope
 import cord.eoeo.momentwo.domain.friend.FriendRepository
-import cord.eoeo.momentwo.ui.BaseViewModel
+import cord.eoeo.momentwo.core.common.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.emptyFlow
 import kotlinx.coroutines.launch
