+++ app/src/main/java/cord/eoeo/momentwo/ui/signup/SignUpContract.kt
@@ -1,9 +1,9 @@
 package cord.eoeo.momentwo.ui.signup
 
 import cord.eoeo.momentwo.data.model.User
-import cord.eoeo.momentwo.ui.UiEffect
-import cord.eoeo.momentwo.ui.UiEvent
-import cord.eoeo.momentwo.ui.UiState
+import cord.eoeo.momentwo.core.common.UiEffect
+import cord.eoeo.momentwo.core.common.UiEvent
+import cord.eoeo.momentwo.core.common.UiState
 
 class SignUpContract {
     data class State(
