+++ app/src/main/java/cord/eoeo/momentwo/ui/signup/SignUpScreen.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import cord.eoeo.momentwo.ui.SIDE_EFFECTS_KEY
+import cord.eoeo.momentwo.core.common.SIDE_EFFECTS_KEY
 import cord.eoeo.momentwo.ui.signup.transformation.DateVisualTransformation
 import cord.eoeo.momentwo.ui.signup.transformation.PhoneVisualTransformation
 import kotlinx.coroutines.CoroutineScope
