+++ app/src/main/java/cord/eoeo/momentwo/ui/signup/SignUpViewModel.kt
@@ -3,7 +3,7 @@ package cord.eoeo.momentwo.ui.signup
 import android.util.Log
 import androidx.lifecycle.viewModelScope
 import cord.eoeo.momentwo.data.signup.SignUpRepository
-import cord.eoeo.momentwo.ui.BaseViewModel
+import cord.eoeo.momentwo.core.common.BaseViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
